Fix issue with ignored pipe negation.
In the parser if the pipe sequence lenght is one (i.e no pipe) then
we do not generate any pipe node in the tree. In case the pipe negation
is there we should generate that node even if the pipe length is 1.

The following code should now work:
! false; echo $?

will return 0
Part of M118-017