[fix][tests] minor correction in UTL3CE004 output
the file 'content' had to be added in testsuite/tests/UTL3CE004/dir_a
in order to allow the testsuite/tests/UTL3CE004/dir_a to be git-committed
(empty dir are not considered by git).

this file was missing in the result of the ls on UTL3CE004/dir_a.

Change-Id: I27f60e2be46a43cffdf400201720c3c187a14b62