Fix build failure on GNU/Linux.
liblua depends on a number of math symbols as well as on libdl
(dlopen et al). This adds those to the list of libs needed to
link. The change affects all non-Windows machines, which is assumed
to be Unix systems. This may break on some Unixen where the math
library is bundled with libc, but shouldn't be much worse than
what we have today.

Change-Id: I0df16157925ff6463f22e0ec543291a7a7667b2c