Fix issue with REPARSE points and move_away function
When a file cannot be deleted because it is in use, we try to move it
away in /tmp/trash. This must be done on the same volume (drive). 
The computation of the volume that actually hosts the file was 
incorrect and resulted in spurious failures 
(for example when c:/gnatmail was a link to g:/gnatmail).
This is now fixed.

Change-Id: I552578b882ffb93578e6f58f641f9d6fbd75e2d4